pyproject.toml
[project]
name = "knuckles"
requires-python = ">=3.11"
dependencies = [
    "requests>=2.31.0",
    "python-dateutil>=2.8.2"
]
version = "0.1.0"

[project.optional-dependencies]
dev = [
    "pip-tools>=7.4.1",
    "ruff>=0.1.2",
    "mypy>=1.4.1",
    "types-requests>=2.31.0.1",
    "types-python-dateutil>=2.8.19.14",

]
tests = [
    "pytest>=7.4.0",
    "responses>=0.23.1",
]

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"


[tool.ruff.lint]
select = ["E", "F", "I001"]

[tool.mypy]
strict = true

[[tool.mypy.overrides]]
module = "tests.*"
disable_error_code = "attr-defined, union-attr"
disallow_untyped_defs = false

[tool.pytest.ini_options]
testpaths = "tests"
